--- src/main/java/ru/yandex/practicum/filmorate/repository/RecommendationRepository.java
@@ -1,9 +1,0 @@
-package ru.yandex.practicum.filmorate.repository;
-
-import ru.yandex.practicum.filmorate.model.Film;
-
-import java.util.List;
-
-public interface RecommendationRepository {
-    List<Film> getByUserIdForLike(Long userId);
-}
